ajout tout

+++ parser.c
@@ -6,11 +6,13 @@
 /*   By: ccommiss <ccommiss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/23 18:31:04 by ccommiss          #+#    #+#             */
-/*   Updated: 2019/02/22 19:45:43 by ccommiss         ###   ########.fr       */
+/*   Updated: 2019/02/22 21:41:48 by ccommiss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "fdf.h" 
+
+
 
 t_fdf	*ft_coord(t_fdf *data, char **file)
 {
@@ -28,7 +30,7 @@ t_fdf	*ft_coord(t_fdf *data, char **file)
 	{
 		if(ft_isdigit(file[0][i]) == 1 && (file[0][i + 1] == ' ' || file[0][i + 1] == '\n'))
 		{
-			data->coord[count] = (float *)malloc(sizeof(float) * 3);
+			data->coord[count] = (int *)malloc(sizeof(int) * 3);
 			data->coord[count][0] = x;
 			data->coord[count][1] = y;
 			if (ft_isdigit(file[0][i - 1]) == 1)
@@ -36,8 +38,8 @@ t_fdf	*ft_coord(t_fdf *data, char **file)
 			else
 				data->coord[count][2] = 0;
 			
-			printf("X-Y-Z VAUT : pt %d [%d - %d - %f] \n", count, x, y, data->coord[count][2]);
-			printf ("DATA %f", data->coord[count][0] );
+			printf("X-Y-Z VAUT : pt %d [%d - %d - %d] \n", count, x, y, data->coord[count][2]);
+			printf ("DATA %d", data->coord[count][0] );
 			x++;
 			count++;
 		}
@@ -84,7 +86,7 @@ void	*ft_analyse(char **file, int fd, t_fdf *data)
 	data->size = data->x_width * data->y_height;
 	printf("WIDTH = %d & HEIGHT = %d && SIZE = %d\n", data->x_width, data->y_height, data->size);
 
-	if (!(data->coord = (float **)malloc(sizeof(float *) * (data->size))))
+	if (!(data->coord = (int **)malloc(sizeof(int *) * (data->size))))
 	 	return (0);
 	printf("ON A MALLOC AVEC SUCCES OU PEUT ETRE PAS\n");
 	data = ft_coord(data, file);
